+++ src/views/quest/components/index.ts
@@ -1,2 +1,2 @@
-export * from './quest-item'
+export * from './quest-list-item'
 export * from './quest-list'
